refactor(Constraints): 기능에 따라 파일 분리
제한 사항 만족 여부를 검사하는 메서드가 담긴 클래스를 기능별로 분리함.

BREAKING CHANGE: 기능에 따른 파일 분리

Before:

입력값의 예외 처리 메서드가 담긴 클래스가 하나의 파일에 모여있다.

After:

각각 담당하는 입력값에 따라 파일을 분리하여, 하나의 파일에 하나의 클래스가 담기도록 변경함.
